2. Интеграционные тесты для проект job4j_cars [#4746]

+++ src/main/java/ru/job4j/cars/repository/car/CarRepository.java
@@ -9,9 +9,9 @@ public interface CarRepository {
 
     Collection<Car> findAll();
 
-    Optional<Car> findById(int id);
+    Optional<Car> findById(Long id);
 
     Optional<Car> save(Car car);
 
-    boolean deleteById(int id);
+    boolean deleteById(Long id);
 }
